Add doors_left list to keep track of user's choices
When I was playing with the original solution program,
I sometimes lost track of which door I had picked and
which would be the door I switched to if I decided to
switch, so I just added code that told the user exactly that.
I also restricted their 2nd choice to be between their
original choice or the door that was not revealed to be
a goat door. This is because I assume once the goat door
is revealed, the user will never choose that door.