Switch basic flake8 linter with Lintly-Flake8
The current flake8 linter only shows console
output, which basically means we have to tell
contributors to look at that (which is hard)
or fix the style ourselves (which takes time
and really shouldn't be our priority).

This new linter I found on GitHub Marketplace
should add reviews to code with style suggestions
based on output from flake8, so that solves that issue.