Fix a couple bugs, clarify instructions
One bug was found in "if guess != wordSet". This is always True since guess is a string and wordSet is a set. Since we only want to decrement guesses when the user guesses wrong,  I changed it to "if guess not in wordSet". Also, I defined score by setting it to 0 (before, it was undefined and would error). I made the hangman function take as a parameter the endword. I clarified the instructions to fit. Also, fixed a bug where you would 'lose' and 'win' at the same time if you guessed the last letter of the word on your last guess.