Update MoveRectProblem.py
We may need to add key up events as well as booleans to be able to move the rectangle. If we don't want a, w, d,s to keep moving the rectangle-we just have to change the indentation. I added 2 extra tabs after keyPressed to prevent the rectangle from moving continuously when a,w,d,s is pressed.